Split web into separate files
